make it a little more safe

+++ clean-homebrew.sh
@@ -1,4 +1,15 @@
 #!/usr/bin/env bash
 
-brew remove --force $(brew list) --ignore-dependencies 
-brew cleanup
+echo "THIS WILL REMOVE ALL BREW PACKAGES"
+
+read -p "Are you sure?" -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	echo "removing all packages"
+	brew remove --force $(brew list) --ignore-dependencies 
+	brew cleanup
+else
+	echo "keeping packages and running brew cleanup"
+	brew cleanup
+fi
